Update length type in "BaseFileStream" struct

+++ include/RED4ext/FileStream.hpp
@@ -14,11 +14,11 @@ struct BaseFileStream
 
     virtual ~BaseFileStream() = 0; // 08
 
-    virtual bool ReadWrite(void* aBuffer, size_t aLength) = 0; // 10
-    virtual LARGE_INTEGER GetPointerPosition() = 0;            // 18
-    virtual LARGE_INTEGER GetSize() = 0;                       // 20
-    virtual bool Seek(LARGE_INTEGER aDistance) = 0;            // 28
-    virtual bool Flush() = 0;                                  // 30
+    virtual bool ReadWrite(void* aBuffer, uint32_t aLength) = 0; // 10
+    virtual LARGE_INTEGER GetPointerPosition() = 0;              // 18
+    virtual LARGE_INTEGER GetSize() = 0;                         // 20
+    virtual bool Seek(LARGE_INTEGER aDistance) = 0;              // 28
+    virtual bool Flush() = 0;                                    // 30
 
     virtual void sub_38() // 38
     {
